Update profiles (including the addition of Babeltrace 2)
I'm removing the `babeltrace-master` profile as the `master` branch of
the Babeltrace repository is the Babeltrace 2 project, not Babeltrace 1.
You can still install Babeltrace 1 with the following profiles:

* `babeltrace-stable-1.2`
* `babeltrace-stable-1.3`
* `babeltrace-stable-1.4`
* `babeltrace-stable-1.5`

This patch adds the Babeltrace 2 project (`babeltrace2`). All the
Babeltrace 2 profiles start with `babeltrace2`.

This patch also updates the `README.adoc` file and the manual pages to
fix mistakes, improve the style, and mention Babeltrace 2.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>