Build babeltrace before LTTng-tools
The LTTng-tools test suite now has a dependency on Babeltrace 2 since it
makes use of custom plug-ins. To ensure the configure step succeeds,
move the build of the Babeltrace projects before LTTng-tools.

Technically, moving babeltrace is unecessary, but it seems preferable to
keep the two projects grouped together in the build order.

Signed-off-by: Jérémie Galarneau <jeremie.galarneau@efficios.com>