Use `importlib.resources` instead of `pkg_resources`
`pkg_resources`, part of the setuptools project, has been deprecated for
a long time, and now, trying vlttng with Python 3.13, it seems like it
just doesn't exist anymore.

Replace all uses of `pkg_resources` with the standard
`importlib.resources`. The latter has a slightly different API, but
it provides the same functionality.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>